Identify when prefixing is unnecessary
A command will get a new logger on each instantiation, the cache in Toolbox is meant to prevent having to rebuild loggers too often,
but in this case we were wrapping the same logger multiple times with a Prefixing Logger on each command execution.

Toolbox wil now check if the most immediate Logger is already a Prefixing Logger of the same prefix.

fixes #4