Fix multi-line alt text issue & clean up
This resolves the issue where multi-line alt texts were not being recognized correctly.

The main culprit was the `PREFIX_PATTERN` regex in CommandMessage and the use `matches()` as opposed to `lookingAt()`.

This also includes some improvements to logging for the purpose of logging only in testing, as well as an execution
context that only runs in testing (for expensive testing-only operations).

Removes Helpers and Anticipated Value along with unnecessarily complicated registry methods.

Interfaces have a more consistent naming scheme now besides ICommandRegistry.

fixes #3